+++ crates/oxc_formatter/src/format.rs
@@ -2,7 +2,7 @@ use oxc_allocator::Vec;
 use oxc_ast::ast::*;
 use oxc_span::GetSpan;
 
-use crate::formatter::{Buffer, CstFormatContext, Format, FormatResult, prelude::*};
+use crate::formatter::{Buffer, Format, FormatResult, prelude::*};
 use crate::{JsFormatContext, write};
 
 pub type JsFormatter<'ast, 'buf> = Formatter<'buf, JsFormatContext<'ast>>;
