+++ crates/oxc_formatter/src/formatter/mod.rs
@@ -25,6 +25,7 @@ mod arguments;
 mod buffer;
 mod builders;
 pub mod comments;
+mod context;
 pub mod diagnostics;
 pub mod format_element;
 mod format_extensions;
@@ -36,6 +37,7 @@ pub mod prelude;
 pub mod printed_tokens;
 pub mod printer;
 pub mod separated;
+mod state;
 mod syntax_element_key;
 mod syntax_node;
 mod syntax_token;
@@ -48,19 +50,20 @@ mod token_text;
 pub mod trivia;
 mod verbatim;
 
+pub use self::arguments::{Argument, Arguments};
 pub use self::formatter::Formatter;
 use self::group_id::UniqueGroupIdBuilder;
 use self::prelude::TagKind;
+pub use self::state::{FormatState, FormatStateSnapshot};
+pub use self::syntax_trivia_piece_comments::SyntaxTriviaPieceComments;
 use std::fmt;
 use std::fmt::{Debug, Display};
 use std::marker::PhantomData;
-
-pub use self::arguments::{Argument, Arguments};
-pub use self::syntax_trivia_piece_comments::SyntaxTriviaPieceComments;
 // use self::builders::syntax_token_cow_slice;
 pub use self::comments::{
     CommentKind, CommentPlacement, CommentStyle, Comments, DecoratedComment, SourceComment,
 };
+pub use self::context::FormatContext;
 pub use self::diagnostics::{ActualStart, FormatError, InvalidDocumentError, PrintError};
 use self::format_element::document::Document;
 #[cfg(debug_assertions)]
@@ -73,7 +76,6 @@ pub use self::text_len::TextLen;
 pub use self::text_range::TextRange;
 pub use self::text_size::TextSize;
 pub use self::token_text::TokenText;
-use crate::context::JsFormatContext;
 // use self::trivia::{format_skipped_token_trivia, format_trimmed_token};
 pub use buffer::{
     Buffer, BufferExtensions, BufferSnapshot, Inspect, PreambleBuffer, RemoveSoftLinesBuffer,
@@ -613,17 +615,6 @@ impl fmt::Display for Expand {
     }
 }
 
-/// Context object storing data relevant when formatting an object.
-pub trait FormatContext<'a> {
-    type Options: FormatOptions;
-
-    /// Returns the formatting options
-    fn source_text(&self) -> &'a str;
-
-    /// Returns the formatting options
-    fn options(&self) -> &Self::Options;
-}
-
 /// Options customizing how the source code should be formatted.
 ///
 /// **Note**: This trait should **only** contain the essential abstractions required for the printing phase.
@@ -645,43 +636,6 @@ pub trait FormatOptions {
 
     /// Derives the print options from the these format options
     fn as_print_options(&self) -> PrinterOptions;
-}
-
-/// The [CstFormatContext] is an extension of the CST unaware [FormatContext] and must be implemented
-/// by every language.
-///
-/// The context customizes the comments formatting and stores the comments of the CST.
-pub trait CstFormatContext<'a>: FormatContext<'a> + Sized {
-    type Style: CommentStyle;
-
-    /// Rule for formatting comments.
-    type CommentRule: FormatRule<SourceComment, Context = Self> + Default;
-
-    /// Returns a reference to the program's comments.
-    fn comments(&self) -> &Comments;
-}
-
-#[derive(Debug, Default, Eq, PartialEq)]
-pub struct SimpleFormatContext {
-    options: SimpleFormatOptions,
-}
-
-impl SimpleFormatContext {
-    pub fn new(options: SimpleFormatOptions) -> Self {
-        Self { options }
-    }
-}
-
-impl<'a> FormatContext<'a> for SimpleFormatContext {
-    type Options = SimpleFormatOptions;
-
-    fn source_text(&self) -> &'a str {
-        ""
-    }
-
-    fn options(&self) -> &Self::Options {
-        &self.options
-    }
 }
 
 #[derive(Debug, Default, Eq, PartialEq, Copy, Clone)]
@@ -1095,144 +1049,6 @@ where
     document.propagate_expand();
 
     Ok(Formatted::new(document, state.into_context()))
-}
-
-/// This structure stores the state that is relevant for the formatting of the whole document.
-///
-/// This structure is different from [crate::Formatter] in that the formatting infrastructure
-/// creates a new [crate::Formatter] for every [crate::write!] call, whereas this structure stays alive
-/// for the whole process of formatting a root with [crate::format!].
-pub struct FormatState<Context> {
-    context: Context,
-    group_id_builder: UniqueGroupIdBuilder,
-    // This is using a RefCell as it only exists in debug mode,
-    // the Formatter is still completely immutable in release builds
-    // #[cfg(debug_assertions)]
-    // pub printed_tokens: PrintedTokens,
-}
-
-impl<Context> std::fmt::Debug for FormatState<Context>
-where
-    Context: std::fmt::Debug,
-{
-    fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
-        f.debug_struct("FormatState").field("context", &self.context).finish()
-    }
-}
-
-impl<Context> FormatState<Context> {
-    /// Creates a new state with the given language specific context
-    pub fn new(context: Context) -> Self {
-        Self {
-            context,
-            group_id_builder: Default::default(),
-            // #[cfg(debug_assertions)]
-            // printed_tokens: Default::default(),
-        }
-    }
-
-    pub fn into_context(self) -> Context {
-        self.context
-    }
-
-    /// Returns the context specifying how to format the current CST
-    pub fn context(&self) -> &Context {
-        &self.context
-    }
-
-    /// Returns a mutable reference to the context
-    pub fn context_mut(&mut self) -> &mut Context {
-        &mut self.context
-    }
-
-    /// Creates a new group id that is unique to this document. The passed debug name is used in the
-    /// [std::fmt::Debug] of the document if this is a debug build.
-    /// The name is unused for production builds and has no meaning on the equality of two group ids.
-    pub fn group_id(&self, debug_name: &'static str) -> GroupId {
-        self.group_id_builder.group_id(debug_name)
-    }
-
-    #[cfg(not(debug_assertions))]
-    #[inline]
-    pub fn track_token(&mut self, _token: &SyntaxToken) {}
-
-    /// Tracks the given token as formatted
-    #[cfg(debug_assertions)]
-    #[inline]
-    pub fn track_token(&mut self, token: &SyntaxToken) {
-        todo!()
-        // self.printed_tokens.track_token(token);
-    }
-
-    #[cfg(not(debug_assertions))]
-    #[inline]
-    pub fn set_token_tracking_disabled(&mut self, _: bool) {}
-
-    /// Disables or enables token tracking for a portion of the code.
-    ///
-    /// It can be useful to disable the token tracking when it is necessary to re-format a node with different parameters.
-    #[cfg(debug_assertions)]
-    pub fn set_token_tracking_disabled(&mut self, enabled: bool) {
-        todo!()
-        // self.printed_tokens.set_disabled(enabled)
-    }
-
-    #[cfg(not(debug_assertions))]
-    #[inline]
-    pub fn is_token_tracking_disabled(&self) -> bool {
-        false
-    }
-
-    /// Returns `true` if token tracking is currently disabled.
-    #[cfg(debug_assertions)]
-    pub fn is_token_tracking_disabled(&self) -> bool {
-        todo!()
-        // self.printed_tokens.is_disabled()
-    }
-
-    #[cfg(not(debug_assertions))]
-    #[inline]
-    pub fn assert_formatted_all_tokens(&self, _root: &SyntaxNode) {}
-
-    /// Asserts in debug builds that all tokens have been printed.
-    #[cfg(debug_assertions)]
-    #[inline]
-    pub fn assert_formatted_all_tokens(&self, root: &SyntaxNode) {
-        todo!()
-        // self.printed_tokens.assert_all_tracked(root);
-    }
-}
-
-impl<'a, Context> FormatState<Context>
-where
-    Context: FormatContext<'a>,
-{
-    pub fn snapshot(&self) -> FormatStateSnapshot {
-        todo!()
-        // FormatStateSnapshot {
-        // #[cfg(debug_assertions)]
-        // printed_tokens: self.printed_tokens.snapshot(),
-        // }
-    }
-
-    pub fn restore_snapshot(&mut self, snapshot: FormatStateSnapshot) {
-        todo!()
-        // let FormatStateSnapshot {
-        // #[cfg(debug_assertions)]
-        // printed_tokens,
-        // } = snapshot;
-
-        // cfg_if::cfg_if! {
-        // if #[cfg(debug_assertions)] {
-        // self.printed_tokens.restore(printed_tokens);
-        // }
-        // }
-    }
-}
-
-pub struct FormatStateSnapshot {
-    // #[cfg(debug_assertions)]
-    // printed_tokens: printed_tokens::PrintedTokensSnapshot,
 }
 
 #[cfg(test)]
